Split yapapi.services to a few separate files
This is not super-final (as some wider refactoring regarding
ServiceRunner is comming next & some things might change).

This commit should change **nothing** in the logic.
